nests gameinprogresscontainer into pregamecontiainer, sim to finalresults
