+++ play/src/App.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { BrowserRouter as Router, Route, Routes } from 'react-router-dom';
 import { ThemeProvider, StyledEngineProvider } from '@mui/material/styles'; // change to mui v5 see CSS Injection Order section of https://mui.com/material-ui/guides/interoperability/
 import { ApiClient, Environment } from '@righton/networking';
-import GameSessionContainer from './containers/GameSessionContainer';
+import PregameContainer from './containers/PregameContainer';
 import Theme from './lib/Theme';
 
 function RedirectToCentralIfMissing() {
@@ -20,7 +20,7 @@ function App() {
           <Routes>
             <Route
               path="/"
-              element={<GameSessionContainer apiClient={apiClient} />}
+              element={<PregameContainer apiClient={apiClient} />}
             />
             <Route element={<RedirectToCentralIfMissing />} />
           </Routes>
