+++ play/src/containers/FinalResultsContainer.tsx
@@ -2,11 +2,11 @@ import React, { useState } from 'react';
 import { useTheme } from '@mui/material/styles';
 import useMediaQuery from '@mui/material/useMediaQuery';
 import { ITeam, GameSessionState } from '@righton/networking';
-import Leaderboard from './finalresults/Leaderboard';
-import Congrats from './finalresults/Congrats';
+import Leaderboard from '../pages/finalresults/Leaderboard';
+import Congrats from '../pages/finalresults/Congrats';
 import { FinalResultsState } from '../lib/PlayModels';
 
-interface FinalResultsProps {
+interface FinalResultsContainerProps {
   teams?: ITeam[];
   currentState: GameSessionState;
   score: number;
@@ -15,18 +15,17 @@ interface FinalResultsProps {
   leader: boolean;
 }
 
-export default function FinalResults({
+export default function FinalResultsContainer({
   teams,
   currentState,
   score,
   selectedAvatar,
   teamId,
   leader,
-}: FinalResultsProps) {
+}: FinalResultsContainerProps) {
   const theme = useTheme();
   const isSmallDevice = useMediaQuery(theme.breakpoints.down('sm'));
   const [finalResultsState, setFinalResultsState] = useState(
-    // eslint-disable-line @typescript-eslint/no-unused-vars
     FinalResultsState.CONGRATS
   );
