+++ play/src/lib/PlayModels.tsx
@@ -60,7 +60,7 @@ export enum PregameState {
   ENTER_GAME_CODE = 'ENTER_GAME_CODE',
   ENTER_NAME = 'ENTER_NAME',
   SELECT_AVATAR = 'SELECT_AVATAR',
-  HOW_TO_PLAY = 'HOW_TO_PLAY',
+  FINISHED = 'FINISHED',
 }
 
 /**
